requirements.txt
anthropic
openai
streamlit
gradio>=4.0.0
python-dotenv
pydantic
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
pillow
requests
python-pptx
